pyproject.toml
[project]
name = "parsers-grpc-service"
version = "1.0.0"
description = "文档解析器 gRPC 微服务 - 支持 PDF/DOCX/PPTX/MD 解析，OCR 文字识别，双层并发优化"
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # 文档解析核心库
    "pdfplumber>=0.10.0",           # PDF 解析（表格检测 + 图像提取）
    "python-docx>=1.0.0",           # DOCX 解析（图文混排支持）
    "python-pptx>=0.6.21",          # PPTX 解析（页面级并发）

    # OCR 引擎
    "paddleocr>=2.10.0",            # PP-OCRv4 server 模型（高准确率）
    "paddlepaddle>=3.0.0",          # PaddlePaddle 深度学习框架
    "pillow>=10.0.0",               # 图像处理
    "numpy>=1.24.0",                # 数值计算

    # gRPC 服务框架
    "grpcio>=1.60.0",               # gRPC 核心库
    "grpcio-tools>=1.60.0",         # Proto 代码生成工具
    "grpcio-health-checking>=1.60.0", # 健康检查服务

    # 环境配置
    "python-dotenv>=1.0.0",         # 环境变量管理
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",                # 单元测试框架
    "pytest-cov>=4.0.0",            # 测试覆盖率
    "pytest-asyncio>=0.21.0",       # 异步测试支持
    "pytest-mock>=3.12.0",          # Mock 支持
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/grpc/generated/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
